chore: Move files around to better places
