+++ packages/core/src/extensions/common-api/stores.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 
-import type { PersistentStorageParams } from "../../common/persistent-storage/create.injectable";
+import type { PersistentStorageParams } from "../../features/persistent-storage/common/create.injectable";
 import type { ExtensionStoreParams } from "../base-extension-store";
 import { BaseExtensionStore as ExtensionStore } from "../base-extension-store";
