+++ packages/core/src/extensions/extension-loader/file-system-provisioner-store/file-system-provisioner-store-injection-token.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectionToken } from "@ogre-tools/injectable";
-import type { MigrationDeclaration } from "../../../common/persistent-storage/migrations.injectable";
+import type { MigrationDeclaration } from "../../../features/persistent-storage/common/migrations.injectable";
 
 export const fileSystemProvisionerStoreMigrationDeclarationInjectionToken = getInjectionToken<MigrationDeclaration>({
   id: "file-system-provisioner-store-migration-declaration",
