+++ packages/core/src/features/cluster/storage/common/migration-token.ts
@@ -4,7 +4,7 @@
  */
 
 import { getInjectionToken } from "@ogre-tools/injectable";
-import type { MigrationDeclaration } from "../../../../common/persistent-storage/migrations.injectable";
+import type { MigrationDeclaration } from "../../../persistent-storage/common/migrations.injectable";
 
 export const clusterStoreMigrationInjectionToken = getInjectionToken<MigrationDeclaration>({
   id: "cluster-store-migration",
