+++ packages/core/src/features/extensions/enabled/common/storage.injectable.ts
@@ -6,9 +6,9 @@ import type { LensExtensionId } from "@k8slens/legacy-extensions";
 import { isDefined } from "@k8slens/utilities";
 import { getInjectable } from "@ogre-tools/injectable";
 import { action, toJS } from "mobx";
-import createPersistentStorageInjectable from "../../../../common/persistent-storage/create.injectable";
-import persistentStorageMigrationsInjectable from "../../../../common/persistent-storage/migrations.injectable";
-import { enabledExtensionsMigrationDeclarationInjectionToken } from "./migrations";
+import createPersistentStorageInjectable from "../../../persistent-storage/common/create.injectable";
+import persistentStorageMigrationsInjectable from "../../../persistent-storage/common/migrations.injectable";
+import { enabledExtensionsMigrationDeclarationInjectionToken } from "../main/migrations";
 import type { LensExtensionState } from "./state.injectable";
 import enabledExtensionsStateInjectable from "./state.injectable";
 import { enabledExtensionsPersistentStorageVersionInitializable } from "./storage-version";
