+++ packages/core/src/features/extensions/enabled/main/migrations.ts
@@ -4,7 +4,7 @@
  */
 
 import { getInjectionToken } from "@ogre-tools/injectable";
-import type { MigrationDeclaration } from "../../../../common/persistent-storage/migrations.injectable";
+import type { MigrationDeclaration } from "../../../persistent-storage/common/migrations.injectable";
 
 export const enabledExtensionsMigrationDeclarationInjectionToken = getInjectionToken<MigrationDeclaration>({
   id: "enabled-extensions-migration-declaration",
