+++ packages/core/src/features/hotbar/storage/common/migrations-token.ts
@@ -4,7 +4,7 @@
  */
 
 import { getInjectionToken } from "@ogre-tools/injectable";
-import type { MigrationDeclaration } from "../../../../common/persistent-storage/migrations.injectable";
+import type { MigrationDeclaration } from "../../../persistent-storage/common/migrations.injectable";
 
 export const hotbarStoreMigrationInjectionToken = getInjectionToken<MigrationDeclaration>({
   id: "hotbar-store-migration-token",
