+++ packages/core/src/features/hotbar/storage/common/storage.injectable.ts
@@ -8,8 +8,8 @@ import { action, comparer } from "mobx";
 import catalogCatalogEntityInjectable from "../../../../common/catalog-entities/general-catalog-entities/implementations/catalog-catalog-entity.injectable";
 import { hotbarStoreMigrationInjectionToken } from "./migrations-token";
 import { defaultHotbarCells } from "./types";
-import createPersistentStorageInjectable from "../../../../common/persistent-storage/create.injectable";
-import persistentStorageMigrationsInjectable from "../../../../common/persistent-storage/migrations.injectable";
+import createPersistentStorageInjectable from "../../../persistent-storage/common/create.injectable";
+import persistentStorageMigrationsInjectable from "../../../persistent-storage/common/migrations.injectable";
 import storeMigrationVersionInjectable from "../../../../common/vars/store-migration-version.injectable";
 import activeHotbarIdInjectable from "./active-id.injectable";
 import createHotbarInjectable from "./create-hotbar.injectable";
