+++ packages/core/src/features/persistent-storage/common/disable-sync.ts
