+++ packages/core/src/features/persistent-storage/main/disable-sync-in-ipc-listener.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { shouldPersistentStorageDisableSyncInIpcListenerInjectionToken } from "../../common/persistent-storage/disable-sync";
+import { shouldPersistentStorageDisableSyncInIpcListenerInjectionToken } from "../common/disable-sync";
 
 const shouldBaseStoreDisableSyncInIpcListenerInjectable = getInjectable({
   id: "should-base-store-disable-sync-in-ipc-listener",
