+++ packages/core/src/features/persistent-storage/main/ipc-channel-prefix.injectable.ts
@@ -3,13 +3,13 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { persistentStorageIpcChannelPrefixesInjectionToken } from "../../common/persistent-storage/channel-prefix";
+import { persistentStorageIpcChannelPrefixesInjectionToken } from "../common/channel-prefix";
 
 const baseStoreIpcChannelPrefixInjectable = getInjectable({
   id: "base-store-ipc-channel-prefix",
   instantiate: () => ({
-    local: "store-sync-renderer",
-    remote: "store-sync-main",
+    local: "store-sync-main",
+    remote: "store-sync-renderer",
   }),
   injectionToken: persistentStorageIpcChannelPrefixesInjectionToken,
 });
